# BUG: 0/frame numeric ops buggy (GH9144)