@Garrett-R 2015-01-20
@jreback, my code is failing because I had assumed that we want 0/0 to be NaN, so implemented that as part of the bugfix, but there's [a test](https://github.com/pydata/pandas/blob/master/pandas/tests/test_frame.py#L5147) that thinks 0/0 should be inf, which seems mistaken.

Should I go ahead and change that test?