@jreback 2015-01-20
see here: http://pandas.pydata.org/pandas-docs/stable/whatsnew.html#id31

also IIRC there was some discussion w.r.t. to 0/0 (but might be in another issue). pls see if you can locate it.