@jreback 2015-01-22
cc @seth-p

iIIRC didn't we have s discussions about this?

what are your thoughts here?