@seth-p 2015-01-22
I agree that ``0/0``, ``0//0``, and ``0.0/0.0``, should be ``NaN`` and not ``inf``.

Is the issue with ``0//0`` that ``NaN`` is actually a float, and so an integer series or DataFrame column cannot contain ``NaN``? So in cases where the result must be an integer (are there such cases?), I'm not sure what we want ``0//0`` to be -- ``0`` seems about as good as anything, I suppose. If there are no such cases, then never mind...

[A completely separate issue is that ``NaN`` is used both for (a) results that are not numbers (e.g. ``0.0/0.0``), and (b) missing values (e.g. ``None``). I could see a reasonable argument for representing those differently.