@jreback 2015-01-22
missing and undefined values are both represented by nan

I know others have different ways of doing this - bottoms line is pandss chose a long time ago to only have a single nan represent both - mainly this a perf issue (to have multiple missing value types)