@Garrett-R 2015-01-22
@seth-p, right, integer types don't support `NaN`, so you have to do one of the following for `0//0`:

 - choose an arbitrary value (NumPy's current behavior)
 - throw an exception (ex: C++ and Python)
 - cast to a float (this is actually what Pandas already does and what they're discussing doing in numpy/numpy#899)

    `assert  (Series([0]) // Series([0])).dtype == np.float_`
    `assert  (Series([0]) // Series([1])).dtype == np.int_`

Given that we're already casting to float, and assuming no one's planning on changing that behavior, we might as well make it `NaN`.  @jreback, cool, I'll get on it!