@jreback 2015-01-22
@Garrett-R the int/float issue is already solved. Pandas handles this correctly. That said I am thinking about this again. 

So now ``1/0`` -> ``inf``, but ``0/0`` -> ``NaN``. Isn't this inconsistent?