@Garrett-R 2015-01-22
@jreback, I believe `1/0` -> `inf` and `0/0` -> `NaN` is conventional (assuming we're casting to floats).  I wanted to try and find it in the IEEE floating point arithmetic document, but I think it's hidden behind an evil paywall.

Anyway, just check out the different languages:

----------------------------------
__C++__

    #include <iostream>
    int main() {
        float x=0.0;
        std::cout << x/x << std::endl;
        std::cout << 1.0/x << std::endl;
    }

outputs

    -nan
    inf

(the negative is irrelevant)

-----------------------------------------

__Julia__

    julia> 0/0
    NaN

    julia> 1/0
    Inf

-----------------------------

__R__

    > 1/0
    [1] NaN
    > 0/0
    [1] Inf