@jreback 2015-01-23
pls add a release note (maybe short example as well), just to make it obvious
put in api change section in 0.16.0