@Garrett-R 2015-02-03
Sorry for the delay.  I've added a release note.

I also modified one more thing.  For modulus, there's an issue:

    In [1]: s = pd.Series([0,1])

    In [2]: s % 0
    Out[2]: 
    0   NaN
    1   NaN
    dtype: float64

    In [3]: 0 % s        # Should give a NaN for 0%0
    Out[3]: 
    0    0
    1    0
    dtype: int64


The modified behavior is:

    In [1]: s = pd.Series([0,1])

    In [2]: s % 0
    Out[2]: 
    0   NaN
    1   NaN
    dtype: float64

    In [3]: 0 % s
    Out[3]: 
    0   NaN
    1     0
    dtype: float64