@jreback 2015-02-10
this should also fix #8445 yes? (if so, pls add that as a test as well), if not, see if its simple to extend to fix.