@Garrett-R 2015-02-11
Thanks for explaining the float issue.  I've undid that change as suggested.

----------------

Yes, it fixes #8445.  I've added that to the commit message.

-------------------

You're not gonna like this but I added more special casing.  It was to correct some new unexpected behavior I found:  `Series([-1,1]) / 0` gave signed infinities, while `Series([-1,1]) // 0` gave only positive infinities.  I believe there's no way to get around the special casing.  I've added the appropriate test.

----------------------

As for the slow division, that was existent before this PR.  But good call on checking into that.  I found that the `_fill_zeros` methods is unnecessary if the result is already a float since then everything will already be good.  Therefore, this PR improves the truediv operation runtime by a factor of 10.


**old behavior**


In [1]: xx = pd.DataFrame(np.random.random((10000, 10000)))

In [2]: %timeit xx / 0
1 loops, best of 3: 1.59 s per loop

**new behavior**

In [1]: xx = pd.DataFrame(np.random.random((10000, 10000)))

In [2]: %timeit xx / 0
10 loops, best of 3: 165 ms per loop