@jreback 2015-02-11
pls add a release note for #8445 (you can simply add it on where you have #9144)

ok on the special casing, sometimes it is unavoidable.

can you add a couple of vbenches (though use a smaller matrix, maybe 1000x1000), the proportion should still be the same; do for float/int results (for floor/div), e.g. add 4 or so (or more if you think are necessary, e.g. maybe for module too). name then consistently and add somewhere in the vbench suite.
help on vbenches here: https://github.com/pydata/pandas/wiki/Performance-Testing.

These are mainly to prevent performance regressions if/when things are changed in the future. Post the results in the top of the PR as well.

thxs