@Garrett-R 2015-02-15
@jreback, how's it looking now?

Also, I've edited my first comment in the PR to include the results of the vbenches (including the ones I added).  Is this what you meant by "the top of the PR"?